requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
supabase>=2.5.0
python-dotenv>=1.0.1
bcrypt>=4.1.3
PyJWT>=2.8.0
pydantic>=2.7.1
httpx>=0.27.0
google-generativeai>=0.5.3
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.22
